pyproject.toml
[project]
name = "hi-paris-research-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu==1.13.2",
    "numpy==2.4.1",
    "openpyxl==3.1.5",
    "pandas==2.3.3",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "scikit-learn==1.8.0",
    "scipy==1.17.0",
    "sentence-transformers==5.2.0",
    "streamlit==1.53.0",
    "torch==2.9.1",
    "tqdm==4.67.1",
    "transformers==4.57.6",
]
